requirements/base.txt
# Core Django packages
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-environ>=0.11.0
python-decouple>=3.8

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# PostGIS
django-geojson>=4.0.0
djangorestframework-gis>=1.0

# Authentication & Authorization
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
PyJWT>=2.8.0

# Celery & Task Queue
celery>=5.3.4
redis>=5.0.1
django-redis>=5.4.0

# Django Channels for WebSocket
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Media & File Processing
Pillow>=10.1.0
django-cloudinary-storage>=0.3.0
python-magic-bin>=0.4.14  # For file type detection on Windows

# API Documentation
drf-spectacular>=0.27.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
phonenumbers>=8.13.26
django-phonenumber-field>=7.3.0

# Validation & Serialization
django-filter>=23.5

# Static files serving
whitenoise>=6.6.0

# Security
django-ratelimit>=4.1.0
cryptography>=41.0.7

# Monitoring & Logging
sentry-sdk>=1.39.1
pythonjsonlogger>=2.0.7  # JSON logging formatter for structured logs

